Fix how vendored assets are imported
this.app is only available for addons who are a direct dependency of the
consuming app.

In the case where the addon is a dependency of another addon, this.app
is not defined.

However, the included function is passed a "parent" which is either the
consuming app or the parent addon.

Seeing as we use ember-colpick as a dependency of an addon which is
itself a dependency of the app, we need this.app fails and we need to
use app instead.